PR 2.3: vectorize collaborator migration
Move the entire Vectorize class (~850 LoC) from
src/pyramids/dataset/ops/vectorize.py into
src/pyramids/dataset/_collaborators.py:Vectorize. Migrated members:

* Public methods: to_feature_collection, translate, cluster, cluster2.
* Private helpers: _band_to_polygon, _extract_values_tiled,
  _extract_values_full, _attach_geometry (staticmethod),
  _nearest_neighbour (staticmethod), _group_neighbours (staticmethod).

Inside the collaborator, every read of Dataset state (band_count,
band_names, columns, crop, crs, get_tile, no_data_value, raster,
read_array, rows) goes through self._ds.<state> via the weakref proxy.
Three call-site fixups required by the migration:

* to_feature_collection's mask path used to bind src to either ``self``
  (the mixin → a Dataset) or to ``self.crop(mask)`` (also a Dataset),
  then call ``src._extract_values_full(...)``. After migration, src
  is now always a Dataset; ``src.vectorize._extract_values_full(...)``
  routes back through the collaborator. ``_attach_geometry(src, ...)``
  takes the Dataset directly so its ``src.get_cell_points`` /
  ``src.get_cell_polygons`` calls hit the Dataset facades from PR 2.1.
* translate constructed the result as ``type(self)(dst, access="write")``.
  Under the proxy, ``type(self._ds)`` returns ``weakref.ProxyType`` —
  use ``self._ds.__class__(dst, ...)`` so the proxy's transparent
  attribute access dispatches to the wrapped Dataset class.
* spatial.py:482 called ``type(self)._nearest_neighbour(...)``; now
  imports Vectorize as _Vectorize and calls
  ``_Vectorize._nearest_neighbour(...)``.

src/pyramids/dataset/dataset.py: drop _VectorizeMixin from the import
alias list and from Dataset's MRO. Add four facades (to_feature_collection,
translate, cluster, cluster2) that thin-forward to self.vectorize.<X>.

src/pyramids/dataset/ops/__init__.py: drop Vectorize import / __all__
entry. src/pyramids/dataset/ops/vectorize.py: keep ``rasterize_features``
free function (still imported by dataset.py); the Vectorize class is
gone.

tests/:

* test_clustering_unit.py: import Vectorize from
  pyramids.dataset._collaborators (the static method
  Vectorize._group_neighbours is still a staticmethod and still callable
  the same way).
* test_dataset_unit.py / test_dataset.py: tests that called
  Dataset._nearest_neighbour or ds._band_to_polygon directly switch to
  Vectorize._nearest_neighbour or ds.vectorize._band_to_polygon.
* test_collaborators.py: move the four vectorize rows from
  FORWARDING_METHODS to FACADE_METHODS.

Full main test suite passes; only the two pre-existing Windows
file-lock flakes in test_dataset_collection.py remain failing.
