+++ src/pyramids/dataset/ops/__init__.py
@@ -4,12 +4,10 @@ from pyramids.dataset.ops.analysis import Analysis
 from pyramids.dataset.ops.band_metadata import BandMetadata
 from pyramids.dataset.ops.io import IO
 from pyramids.dataset.ops.spatial import Spatial
-from pyramids.dataset.ops.vectorize import Vectorize
 
 __all__ = [
     "Analysis",
     "BandMetadata",
     "IO",
     "Spatial",
-    "Vectorize",
 ]
